requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
supabase==2.0.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
